Correction after comments and removed all file
