+++ start_scripts/start_sim.sh
@@ -42,8 +42,7 @@ tmux send-keys "rosrun vehicle_sim sim.py" C-m
 tmux select-pane -t 3
 tmux send-keys "cd '$ABS_SCRIPT_DIR/../autonomous_software_pkg/src'" C-m
 tmux send-keys "source $setup_bash_path" C-m
-# tmux send-keys "rosrun autonomous_software_pkg map_publisher.py" C-m
-tmux send-keys "rosrun autonomous_software_pkg map_publisher_2.py _map_file_name:=MapSanMateoP1.kml" C-m
+tmux send-keys "rosrun autonomous_software_pkg map_publisher.py _map_file_name:=MapSanMateoP1.kml" C-m
 
 # waypoint publisher
 tmux select-pane -t 4
